Simplify Town entity to townName string
