+++ Backend/src/test/java/com/example/backend/CompanyServiceTests.java
@@ -43,9 +43,9 @@ public class CompanyServiceTests
     public void getAllCompanies_ReturnsCompanies() throws AuthenticationException
     {
         List<Company> companies = List.of(
-                new Company(Long.valueOf(1), "name", "domain", 'a', "March", "Croatia", 10000, "Address", "example.com", false, "opis"),
-                new Company(Long.valueOf(2), "name", "domain", 'a', "March", "Croatia", 10000, "Address", "example.com", false, "opis"),
-                new Company(Long.valueOf(3), "name", "domain", 'a', "March", "Croatia", 10000, "Address", "example.com", false, "opis"));
+                new Company(Long.valueOf(1), "name", "domain", 'a', "March", "Croatia", "Zagreb", "Address", "example.com", false, "opis"),
+                new Company(Long.valueOf(2), "name", "domain", 'a', "March", "Croatia", "Zagreb", "Address", "example.com", false, "opis"),
+                new Company(Long.valueOf(3), "name", "domain", 'a', "March", "Croatia", "Zagreb", "Address", "example.com", false, "opis"));
         Mockito.when(companyRepository.findAll()).thenReturn(companies);
 
         List<Company> result = companyService.getAllCompanies(mockUser(AUTHORITY.ADMINISTRATOR));
@@ -63,7 +63,7 @@ public class CompanyServiceTests
     @Test
     public void getCompany_IfUserIsAdmin_ReturnsCompany() throws AuthenticationException
     {
-        Company company = new Company(Long.valueOf(1), "name", "domain", 'a', "March", "Croatia", 10000, "Address", "example.com", false, "opis");
+        Company company = new Company(Long.valueOf(1), "name", "domain", 'a', "March", "Croatia", "Zagreb", "Address", "example.com", false, "opis");
         Mockito.when(companyRepository.findById(Mockito.any())).thenReturn(Optional.of(company));
         Company result = companyService.getCompany(mockUser(AUTHORITY.ADMINISTRATOR), Long.valueOf(1));
         assertThat(result).isSameAs(company);
@@ -78,7 +78,7 @@ public class CompanyServiceTests
                 'a',
                 "February",
                 "Croatia",
-                10000,
+                "Zagreb",
                 "Address",
                 "www.example.com",
                 false,
@@ -89,7 +89,7 @@ public class CompanyServiceTests
                 'a',
                 "February",
                 "Croatia",
-                10000,
+                "Zagreb",
                 "Address",
                 "www.example.com",
                 false,
